20w24c
Started convertion between LWRP to URP, this enables me to add my own rendering techniques. Created a shader for the interactables when hidden behind an object. Started creating an inticator of when the player is hovering over an interactable. B9) ray cast does not update for a specific layer under some cases